pyproject.toml
[tool.poetry]
name = "private-assistant-comms-bridge"
version = "0.1.0"
description = "Comms Bridge is an open-source endpoint designed to facilitate speech interaction within smart assistant systems, optimized for operation on edge devices."
authors = ["stkr22 <stkr22@github.com>"]
readme = "README.md"
license = "MIT"
include = ["src"]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
librosa = "^0.10.1"
paho-mqtt = "^2.0.0"
openwakeword = "^0.6.0"
httpx = "^0.27.0"
pyyaml = "^6.0.1"
pydantic = "^2.6.4"
typer = "^0.11.1"
sounddevice = "^0.4.6"
torchaudio = "^2.2.2"
speexdsp-ns = "^0.1.2"

[tool.poetry.group.dev.dependencies]
ruff = ">=0.3.4"
mypy = ">=1.8.0"
pytest = ">=8.1.1"
types-pyyaml = "^6.0.12.20240311"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I", "UP"]
ignore = ["E203"]
